Added automatic handling domain boundaries, see full description.
Several modifications have been grouped in this commit:

- The position of the boundaries between domains is calculated automatically.
Within each zone (convective, radiative) it is chosen so that each domain
covers the same interval in log(pressure). As a consequence, the parameter xif
can no longer be modified by the user. Models calculated with a previous version,
when used as input, will be recalculated to verify the new prescription.

- New mapping interpolator. Now it is possible to change the number of domains
of an input model.

- Core convection is automatically detected and the number of domains in the core
is chosen automatically depending on the interval in log(p).
Parameter conv is no longer modifiable. There are 2 new global parameters:

	- core_convec (default 1). To disable core convection, set core_convec=0
	- min_core_size (default 0.01). If the size of the core (in fraction of the
		polar radius) is less than this value, it is removed.
		
This parameters can be modified in config/star.cfg or in the command line 
"-parameter value". They are only used for the current calculation and they 
are not saved with the model.

Other minor changes:

- Compiler settings in make.inc, are used also to compile the Houdek library.
- Makefile: Removed extra space after -I and -L.
- Other small changes and bug fixes.



git-svn-id: http://ester-project.googlecode.com/svn/trunk@54 dc0d849c-13d9-9696-9681-8a32c59e610f