Point le 20/12/2017
entropy rescaled in eos_ideal.cpp, eos_idealrad.cpp, eos_opal.cpp

We investigated the construction of s=cte zones on the M=2Msun model.

The model seems to first converge with a zone around r=0.991-0.994
but when the s-profile is almost flat a jump appears and a very thin
domain appears (noise on schwarz ?).

We now attempt to introduce convection in the last domain following
physics...