Point le 22/12/2017 , 21h04
star_map.cpp: We imposed 2 interfaces to the model, ie 1 CC and 1 CZ. The
CZ is determined from the first vanishing of BV2 after the radiative zone
that is above the CC.  For a 2 Msun model the CZ happens to be split into
3 domains. We have tried to set convection up to surface but leads to
numerical oscillations.  We have arbitrarily changed the last domain to
a radiative domain to accommodate an entropy drop. Then the convergence
reappears to 1e-6. Entropy is rather nice in the file entropie.txt but
burk in the M2.h5 file (a pb there which likely impacts the computation
of other models at lower masses).

star1d_solvers.cpp: small modif to impose last_domain to be RADIATIVE.
All the interface conditions have been checked and seem to be OK. The role
of the condition L=4*pi*r^2*F is not clear except at the surface of the core.

The code runs with a M=2 Msun, giving Teff=8459.7 K
but the change of sign of BV2 remains very noisy.

Next steps:
1/ check the output file for s
2/ the MLT might be the solution for surface convection as it allows
for an entropy drop and washes out all the opacities problems that plague cold
surface layers.