Little bug solved in dev-michel+review of interface conditions
We solved the bug of random start in branch dev-michel (delta was not
initialized in star1d.cpp). For that ester_err("ERROR tracing...");
was introduced in solver.cpp

Some improvement of output visu

interface condition revised: DT continuous at top of CZ.
but left discontiunous at bottom of CZ.

Comments in solve_map

Still a bug on the entropy profile.