New version entropy diffusion in convection zones
It is derived from the version with ds=0 in CZ.
Energy conservation is verified much better and
changes in Pe do not yield convergence problem.

However inspection of the flux show numerical wiggles in the convective
layer and a one point singularity at the bottom of the layer.

To be improved.....