New version with entropy diffusion
Continuity condition at top CZ are now set to
flux=L/(4pi*r^2*khi) ==> better automatism on radiative domains

Now the file_solution.h5 can be used as input and converges.

BUT

The flux continuity is violated by 2% and the Energy test by the same amount.

In the core we keep Ds=0