Add an entropy dependent Peclet
We add a Peclet function of grads : Pe=Pe0/2*(1-th(ds/ds0))
We also make Pe0 an adjustable parameter that evolves to
the targeted Peclet.
We could compute a "sun" with Pe=600 (ds0=0.2, surff=10),
but the convergence is very slow and the solution (at Pe=600)
is not smooth in the outer layers (sawtooth oscillations).
We note that the Peclet does not vanish at the surface
which probably generates the trouble...