python/ester.in and physics/composition.cpp modified
to comply with some changes introduced in ubuntu24.04 we modified
ester.in: problems seem to come from the wrapper SWIG.
The correction added however need the path
ester_root_dir/local/lib/python3/dist-packages/ester
be included in the PYTHONPATH env. variable

composition.cpp: modified also to take into account that the rootdir of
ESTER might not be $HOME/ester. To be customized by author.