Change in nuc.name simple
On November 4th I moved the XCNO=Z*(C+N) to XCNO=Z*(O+N). However this move was
too drastic and corrupted the backward compatibility of the code. So now there
are 2 options when running: -nuc simple_CN and -nuc simple_ON and old files are
assumed simple_CN.